Group post-processing to a "tagging" category
